% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_similarity_network.R
\name{analyze_similarity_network}
\alias{analyze_similarity_network}
\title{Analyzes immuneNET similarity network}
\usage{
analyze_similarity_network(similarity_network)
}
\arguments{
\item{similarity_network}{immuneREF output layer}
}
\value{
List of network features
}
\description{
Analyzes immuneNET similarity network
}
\examples{
\dontrun{
analyze_similarity_network(similarity_network)
}
}
