% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_characteristics.R
\name{calc_characteristics}
\alias{calc_characteristics}
\title{CALCULATES IMMUNEREF LAYER CHARACTERISTICS}
\usage{
calc_characteristics(
  repertoire_df,
  models = list(svm_public_mm = svm_model_public),
  species,
  receptor,
  chain,
  identifier_rep,
  vdj_list = list_germline_genes,
  dictionary = dictionary_counts,
  architecture_depth = 1,
  aa_range = c(8:20)
)
}
\arguments{
\item{repertoire_df}{Input repertoire (AIRR standard column naming)}

\item{models}{List of ML models compatible with kebabs::predict() function}

\item{species}{Species information of input repertoire ("hs" or "mm")}

\item{receptor}{Receptor information of input repertoire ("ig" or "tr")}

\item{chain}{Chain information of input repertoire ("h","l","b" or "a")}

\item{identifier_rep}{ID-name for input repertoire}

\item{vdj_list}{Reference list for germline gene naming (provided dataset: list_germline_genes)}

\item{dictionary}{Dictionary for kmer-occurrence search (provided: dictionary_counts)}

\item{architecture_depth}{maximal LD for network calculation (a network is calculated for each distance 1:architecture_depth)}

\item{aa_range}{range of sequence lengths for which amino acid frequencies should be evaluated}
}
\value{
List of repertoire layers
}
\description{
CALCULATES IMMUNEREF LAYER CHARACTERISTICS
}
\examples{
\dontrun{
calc_characteristics(repertoire_df,models=list(svm_public_mm=svm_model_public),species,receptor,chain,identifier_rep,vdj_list=list_germline_genes,dictionary=dictionary_counts)
}
}
