% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_similarities.R
\name{calculate_similarities}
\alias{calculate_similarities}
\title{Calculates Similarity Layers BETWEEN REPERTOIRES}
\usage{
calculate_similarities(
  repertoires_analyzed,
  overlap_layer = "",
  vdj_vj_weights = c(1, 1, 1, 0),
  convergence = "overlap",
  cor_methods = c(Diversity = "pearson", AAfreq = "pearson", Architecture = "pearson",
    Convergence = "pearson", VDJ_usage = "pearson", `k-mers` = "pearson")
)
}
\arguments{
\item{repertoires_analyzed}{immuneREF analyzed repertoires}

\item{overlap_layer}{overlap_layer if calculated_previously and convergence method is 'overlap'}

\item{vdj_vj_weights}{How VDJ usage should be treated for similarity calculation (V,D,J,VJusage on/off)}

\item{convergence}{Which convergence measure should be used overlap or immunosignatures (Default: overlap)}

\item{cor_methods}{vector of lengths 6 determining which correlation method should be used for each layer.}
}
\value{
list_single_layers
}
\description{
Calculates Similarity Layers BETWEEN REPERTOIRES
}
\examples{
\dontrun{
calculate_similarities(repertoires_analyzed,vdj_vj_weights=c(1,0,1,0),convergence="overlap",overlap_layer="",cor_methods=c("Diversity"="pearson","AAfreq"='pearson',"Architecture"="pearson","Convergence"='pearson',"VDJ_usage"="pearson","k-mers"='pearson'))
}
}
