% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_check.R
\name{compatibility_check}
\alias{compatibility_check}
\title{Tests immuneREF-compatibility of repertoire}
\usage{
compatibility_check(repertoire, species, receptor)
}
\arguments{
\item{repertoire}{Input immune repertoire}

\item{species}{Species of input repertoire}

\item{receptor}{Receptor chain of immune repertoire ("igh" or "trb")}
}
\value{
Outputs ERRORs when encountered and returns number of errors.
}
\description{
Tests immuneREF-compatibility of repertoire
}
\examples{
\dontrun{
compatibility_check(repertoire, "mm", "igh")
}
}
