% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_layers.R
\name{condense_layers}
\alias{condense_layers}
\title{Condense single IMMUNENET LAYERS}
\usage{
condense_layers(list_single_layers, weights, method = "standard")
}
\arguments{
\item{list_single_layers}{List of calculated immuneREF similarity layers}

\item{weights}{Weights for each layer}

\item{method}{Method for condensing (so far only "standard" is available)}
}
\value{
output matrix
}
\description{
Condense single IMMUNENET LAYERS
}
\examples{
\dontrun{
condense_layers(list_single_layers, weights, method="spearman")
}
}
