% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_counts}
\alias{dictionary_counts}
\title{List of gapped kmer patterns.}
\format{
A list of with 4 entries :
\describe{
  \item{nt}{List of pairs of nucleotide 3-mers}
  \item{aa}{List of pairs of amino acid  1-mers}
  \item{gap_nt}{Numeric vector containing allowed gap sizes for nucleotide sequences}
  \item{gap_aa}{Numeric vector containing allowed gap sizes for amino acid sequences}
}
}
\usage{
dictionary_counts
}
\description{
A list that contains templates for the counting of gapped-kmers
with parameters k = 1, m <= 1 for amino acid
and parameters k = 3, m <= 3 for nucleotide sequences.
}
\keyword{datasets}
