% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneREF_quickstart.R
\name{immuneREF_quickstart}
\alias{immuneREF_quickstart}
\title{Quickstart function to calculate immuneREF layers from list of reperotires}
\usage{
immuneREF_quickstart(repertoire_list, convergence_layer = "overlap")
}
\arguments{
\item{repertoire_list}{List of AIRR-compliant repertoires.}

\item{convergence_layer}{Determines whether Convergence layer is calculated using Overlap or Immunosignatures}
}
\value{
List of immuneREF similarity layers
}
\description{
Quickstart function to calculate immuneREF layers from list of reperotires
}
\examples{
\dontrun{
immuneREF_quickstart(tutorial_repertoires)
}
}
