% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mixcr_output.R
\name{load_mixcr_output}
\alias{load_mixcr_output}
\title{Loads output from MiXCR}
\usage{
load_mixcr_output(filename, AA_length_of_seqs = 0)
}
\arguments{
\item{filename}{Address of file to be loaded}

\item{AA_length_of_seqs}{if !=0 will only include sequences of one specified length}
}
\value{
Processed repertoires
}
\description{
Loads output from MiXCR
}
\examples{
\dontrun{
load_mixcr_output(filename,AA_length_of_seqs)
}
}
