% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cormat.R
\name{make_cormat}
\alias{make_cormat}
\title{Calculates similarities between processed repertoires}
\usage{
make_cormat(
  repertoires_analyzed,
  weights_VDJ = c(1, 0, 1, 0),
  weights_overall = c(1, 1, 1, 1, 1, 1),
  correlation_method = "pearson",
  arch_method = "all"
)
}
\arguments{
\item{repertoires_analyzed}{Output of feature analysis}

\item{weights_VDJ}{V,D,J,VJ weights for similarity calculation}

\item{weights_overall}{Layer weights for similarity calculation}

\item{correlation_method}{Method for correlation calculations}

\item{arch_method}{How architecture layer should be calculated (Default="all")}
}
\value{
Matrix containing similarity of each repertoire pair
}
\description{
Calculates similarities between processed repertoires
}
\examples{
\dontrun{
make_cormat(repertoires_analyzed,weights_VDJ=c(1,1,1,0),
weights_overall=c(1,1,1,1,1,1),correlation_method='pearson',
arch_method="all")
}
}
