% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kmer_dictionary.R
\name{make_kmer_dictionary}
\alias{make_kmer_dictionary}
\title{Preparatory step for kmer analysis makes kmer dictionary}
\usage{
make_kmer_dictionary(
  k_nt,
  k_aa,
  gap_size_nt = c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
  gap_size_aa = c(0, 1, 2, 3)
)
}
\arguments{
\item{k_nt}{length of nucleotide pattern}

\item{k_aa}{length of amino acid pattern}

\item{gap_size_nt}{gap size for nt pattern}

\item{gap_size_aa}{gap size for aa pattern}
}
\value{
Dictionary of gapped kmers
}
\description{
Preparatory step for kmer analysis makes kmer dictionary
}
\examples{
\dontrun{
make_kmer_dictionary(k_nt,k_aa,
gap_size_nt=c(0,1,2,3,4,5,6,7,8,9,10),
gap_size_aa=c(0,1,2,3))
}
}
