% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_local_similarity.R
\name{print_local_similarity}
\alias{print_local_similarity}
\title{Prints violin plots of local similarity values per layer, across categories}
\usage{
print_local_similarity(
  list_similarity_matrices,
  categories_list,
  path_figure = "figures"
)
}
\arguments{
\item{list_similarity_matrices}{list of immuneREF layers}

\item{categories_list}{List with entries 'categories' and 'color' and 'subset' for plot structure}

\item{path_figure}{Path where plots should be stored}
}
\value{
TRUE
}
\description{
Prints violin plots of local similarity values per layer, across categories
}
\examples{
\dontrun{
print_local_similarity(list_similarity_matrices,categories_list,path_figure="figures")
}
}
