% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_repertoire_comparison.R
\name{print_repertoire_comparison}
\alias{print_repertoire_comparison}
\title{Outputs repertoire comparison plot}
\usage{
print_repertoire_comparison(
  list_repertoires,
  mode = "list",
  aa_freq_length = 14,
  chain = "H",
  path_figure = "",
  name_plots = ""
)
}
\arguments{
\item{list_repertoires}{List of AIRR-compliant repertoires.}

\item{mode}{Input format "list" or "single" repertoire}

\item{aa_freq_length}{Length for which AA-frequency plot should be output}

\item{chain}{Chain of receptor (Default="H", alternatives not implemented)}

\item{path_figure}{Path where figures should be saved}

\item{name_plots}{Naming for plot}
}
\value{
TRUE
}
\description{
Outputs repertoire comparison plot
}
\examples{
\dontrun{
print_repertoire_comparison(list_repertoires)
}
}
