% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_repertoire_radar.R
\name{print_repertoire_radar}
\alias{print_repertoire_radar}
\title{Prints violin plots of similarity values per layer, across categories}
\usage{
print_repertoire_radar(
  list_similarity_matrices,
  to_compare,
  path_figure,
  name_plot
)
}
\arguments{
\item{list_similarity_matrices}{list of immuneREF layers}

\item{to_compare}{List with entries 'roi' (repertoires of interest), 'roi names', 'ref',' 'plot_names' and 'colors'}

\item{path_figure}{Path where plots should be stored}

\item{name_plot}{Name of plot}
}
\value{
TRUE
}
\description{
Prints violin plots of similarity values per layer, across categories
}
\examples{
\dontrun{
print_repertoire_radar(list_similarity_matrices,to_compare,path_figure,name_plot)
}
}
