% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_overlap.R
\name{repertoire_overlap}
\alias{repertoire_overlap}
\title{CALCULATES PAIRWISE OVERLAP BETWEEN REPERTOIRES}
\usage{
repertoire_overlap(list_simulated_repertoires, basis = "CDR3_aa")
}
\arguments{
\item{list_simulated_repertoires}{list of repertoires to be analyzed}

\item{basis}{Basis on which overlap is calculated (default="CDR3_aa)}
}
\value{
overlap_stat
}
\description{
CALCULATES PAIRWISE OVERLAP BETWEEN REPERTOIRES
}
\examples{
\dontrun{
repertoire_overlap(list_simulated_repertoires=tutorial_repertoires,basis="CDR3_aa")
}
}
