% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_input_repertoires.R
\name{subset_input_repertoires}
\alias{subset_input_repertoires}
\title{Subsets list of input repertoires to desired nb of sequences.}
\usage{
subset_input_repertoires(list_repertoires, subset_size = 10000, random = FALSE)
}
\arguments{
\item{list_repertoires}{list of input repertoires}

\item{subset_size}{maximal number of sequences per repertoire (default 10'000)}

\item{random}{If TRUE subsetting is done by random subsampling, else top x clones are picked (by frequency)}
}
\value{
List of subsampled repertoires features
}
\description{
Subsets list of input repertoires to desired nb of sequences.
}
\examples{
\dontrun{
subset_input_repertoires(tutorial_repertoires,100, FALSE)
}
}
