% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svm_model_public}
\alias{svm_model_public}
\title{A pretrained Support Vector Machine model for the prediction of public clones}
\format{
A kebabs SVM model.
\describe{
  \item{svm_model_public}{Trained on 109'568 nucleotide sequences with gapped-kmers k = 3,m = 1}
}
}
\source{
\url{https://doi.org/10.4049/jimmunol.1700594}
}
\usage{
svm_model_public
}
\description{
An SVM model for the prediction of public clones trained on murine B-cell data.
As published by Greiff et al. (2017)
}
\keyword{datasets}
