% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tutorial_repertoires}
\alias{tutorial_repertoires}
\title{Tutorial repertoires for code examples}
\format{
A list containing 4 dataframes containing 10'000 sequences each.
\describe{
  \item{Repertoire 1}{sim_mm_ig_h_2_0__A}
  \item{Repertoire 2}{sim_mm_ig_h_4_0__A}
  \item{Repertoire 3}{sim_hs_ig_h_2_0__A}
  \item{Repertoire 4}{sim_hs_ig_h_4_0__A}
}
}
\usage{
tutorial_repertoires
}
\description{
A list containing five simulated repertoires using the immuneSIM package.
}
\keyword{datasets}
